utils: add credentialExists() function
This will allow us to detect if a credential exists before we try
to use it and will be able to influence some of our logic about when
we do certain actions.

For example, if there is no 'oscontainer-secret' to allow us to push
the container to quay.io then let's not try.