jobs/bump-lockfile: send lockfile updates to aarch64 builder
It turns out that we were properly updating the aarch64 lockfile
in the fetch but we weren't actually sending that to the remote
builder when doing the build/test pieces. This meant the aarch64
tests in bump-lockfile were essentially always testing what was
in git and not the updated lockfiles. Let's forward the changes
to the remote builder.

I had to use patch here because there isn't an easy way to forward
arbitrary files via gangplank so this was just a way to do it.

Also note that originally I tried a different way to apply the patch
like `git apply <(base64 -d <<< 'abcdefg..')` but that fails to render
properly because < and > get turned into &lt and &rt because we are
using html templating on the backend in gangplank.

I also took this as an opportunity to break out and call `cosa fetch`
and `cosa build` explicitly in the spec rather than using
`build_artifacts: [base]`, which is a little more opaque to me.